requirements.txt
python-gitlab
PyYAML
jinja2
jinja2_base64_filters
xmlplain
